oppdatere filstruktur
