--- AuthenticationTest/BraceletTest/BraceletTest.java
@@ -1,21 +1,0 @@
-package AuthenticationTest.BraceletTest;
-
-public class BraceletTest {
-    
-
-public static void main(String[] args) {
-    System.out.println("Device found!   ID 0101#ab");
-    
-    String patientName = "Karolina Gajos";
-    int patientAge = 94;
-    String patientAllergies = "Peanuts, Shellfish, Common sense";
-    String patientAddress = "123 Oslo Street, Oslo, Norway";
-    
-    System.out.println("Patient Information:");
-    System.out.println("Name: " + patientName);
-    System.out.println("Age: " + patientAge);
-    System.out.println("Allergies: " + patientAllergies);
-    System.out.println("Address: " + patientAddress);
-}
-
-}
